+++ backend/sse-strategy-matchers/src/main/java/io/featurehub/strategies/matchers/DateArrayMatcher.java
@@ -1,6 +1,6 @@
-package io.featurehub.edge.strategies.matchers;
+package io.featurehub.strategies.matchers;
 
-import io.featurehub.mr.model.RolloutStrategyAttribute;
+import io.featurehub.sse.model.RolloutStrategyAttribute;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
