+++ backend/sse-strategy-matchers/src/main/java/io/featurehub/strategies/percentage/Murmur3_32HashFunction.java
@@ -1,4 +1,4 @@
-package io.featurehub.edge.strategies;
+package io.featurehub.strategies.percentage;
 
 /*
  * Copyright (C) 2011 The Guava Authors
