+++ backend/sse-strategy-matchers/src/main/java/io/featurehub/strategies/percentage/PercentageCalculator.java
@@ -1,4 +1,4 @@
-package io.featurehub.edge.strategies;
+package io.featurehub.strategies.percentage;
 
 public interface PercentageCalculator {
   int determineClientPercentage(String userKey, String id);
